requirements.txt
# Requirements for Qwen3-VL-8B-Instruct fine-tuning script
# Install with: pip install -r requirements.txt

# Core ML libraries
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.45.0
datasets>=2.14.0
tokenizers>=0.19.0
accelerate>=0.34.0
peft>=0.12.0

# Quantization and optimization
bitsandbytes>=0.41.0
scipy>=1.10.0
numpy>=1.24.0

# Utilities
tqdm>=4.64.0
psutil>=5.9.0
pyyaml>=6.0
requests>=2.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.15.0
tensorboard>=2.13.0

# System utilities
pyyaml
requests